requirements.txt
boto3==1.34.0
pyspark==3.5.0
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.13.0
great-expectations==0.18.8
pandas==2.1.4
pyarrow==14.0.1
pytest==7.4.3
python-dotenv==1.0.0
